backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
schedule
gunicorn
python-dotenv
google-generativeai
